Use PhantomJS to render Phabricator project board
Phabricator now prints an empty board in HTML, and some JS initial
data content. The goal is to allow live update of the dashboard.

`maniphest.search` project attachment doesn't include the column,
so to use a scraper is still a valid solution.

As we don't have the HTML content of the tasks anymore already
rendered (it's at the bottom, ready to be injected with the
JX.Stratcom.mergeData method), we use PhantomJS, an headless
webkit client to get the rendered HTML content.

The new class HtmlPage takes care of initialize and configure
such client, and get the content.

This change pulls Composer dependencies, and as such, requires
a PHAR build code update, to handle the subdirectories.