Tidy repository file structure (part 1)
Composer configuration file is currently in src/, a unusual position
breaking some workflows.

There is a extranous todeploy symlink between the one in bin/ and the
one in src/, we remove the second.

This part 1 focus on everytinh excepted the code, which will wait we
load our classes through Composer in PSR-4 mode.